+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - Documentation
   - [Todo] Add _cljdoc_ badge
   - Reformat docstring for _cljdoc_
-  - [Todo] Add documentation page with context and config keys
+  - Add documentation page with context and config keys
 
 
 ## 0.6.1-0.1.0 / 2018-October-10