+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   - [Todo] Sync with https://w3c.github.io/distributed-tracing/report-trace-context.html
 
 
-## [WIP] 0.6.2-0.2.0 / 2020-October-??
+## 0.6.2-0.2.0-alpha1 / 2020-October-03
 
 - Upgrade bract.core to 0.6.2
 - Support HTTP `POST` method (with custom body) in `/ping` middleware